requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
pinecone-client==2.2.4
requests==2.31.0
python-dotenv==1.0.0
gitpython==3.1.40
streamlit==1.28.2
openai==1.12.0
beautifulsoup4==4.12.3
langchain==0.0.350
python-markdown-math==0.8
tiktoken==0.6.0
numpy==1.26.4
pandas==2.2.1
pydantic==1.10.13 